+++ src/pylibreconcile/known_state/__init__.py
@@ -1,5 +1,6 @@
 from .aws import AWSS3KnownStateHandler
 from .azure import AzureStorageKnownStateHandler
+from .boltdb import BoltDBKnownStateHandler
 from .json_local import LocalJSONKnownStateHandler
 from .protocol import KnownStateHandler
 from .sqlite import SQLiteKnownStateHandler
@@ -8,6 +9,7 @@ from .yaml_local import LocalYAMLKnownStateHandler
 __all__ = [
     "AWSS3KnownStateHandler",
     "AzureStorageKnownStateHandler",
+    "BoltDBKnownStateHandler",
     "KnownStateHandler",
     "LocalJSONKnownStateHandler",
     "LocalYAMLKnownStateHandler",